requirements.txt
Django==1.6.5
Markdown==2.4.1
Pillow==2.4.0
Pygments==1.6
South==0.8.4
Unidecode==0.04.16
bpython==0.13
dj-database-url==0.3.0
dj-static==0.0.5
django-annoying==0.8.0
django-facebook==6.0.0
django-toolbelt==0.0.1
gunicorn==19.0.0
hamlpy==0.82.2
psycopg2==2.5.3
pystache==0.5.4
static==1.0.2
wsgiref==0.1.2
